requirements.txt
# -----------------------------
# 🌐 Core Framework
# -----------------------------
Flask>=3.0
Werkzeug>=3.0
Jinja2>=3.1

# -----------------------------
# 🗄️ Banco de Dados e Migrações
# -----------------------------
Flask-SQLAlchemy>=3.1
Flask-Migrate>=4.0
alembic>=1.13
PyMySQL>=1.1

# -----------------------------
# 🔒 Autenticação e Segurança
# -----------------------------
Flask-Login>=0.6
Flask-WTF>=1.2
email-validator>=2.1
cryptography>=42.0

# -----------------------------
# 📧 E-mail
# -----------------------------
Flask-Mail>=0.9

# -----------------------------
# ⚙️ Utilitários
# -----------------------------
python-dotenv>=1.0
